pyproject.toml
[project]
name = "recidiviz"
version = "0.1.0"
description = "Recidiviz data platform for criminal justice reform"
requires-python = ">=3.11, <3.12"
dependencies = [
    "annotated-types",
    # TODO(#30733): Remove the <2.0.0 pin
    "numpy<2.0.0",
    "us",
    # TODO(#60465): Un-pin when airflow environment is upgraded to a compatible version
    "dateparser==1.2.0",
    "pandas",
    "more-itertools",
    # TODO(#43270): Un-pin opentelemetry-* packages once resolved
    "opentelemetry-api<1.34.0",
    "opentelemetry-sdk<1.34.0",
    "opentelemetry-exporter-gcp-monitoring",
    "opentelemetry-exporter-gcp-trace",
    "opentelemetry-resourcedetector-gcp",
    "opentelemetry-instrumentation-flask",
    "opentelemetry-instrumentation-grpc",
    "opentelemetry-instrumentation-redis",
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-semantic-conventions",
    "opentelemetry-propagator-gcp",
    # TODO(#18145): Remove pin once migration 94563473c612 is fixed to not
    # pass primary_key=True to op.add_column(). Alembic 1.18.2+ emits
    # PRIMARY KEY inline in ADD COLUMN, breaking tables with existing PKs.
    # Using the same issue as the sqlalchemy pin below, because updating that
    # will probably need to be done in tandem with alembic.
    "alembic<1.18.2",
    # TODO(#18145): Remove the <2.0.0 pin
    "sqlalchemy>=1.4.25,<2.0.0",
    "flask[async]>=2.0.0",
    "Flask-Limiter",
    "Flask-SQLAlchemy-Session",
    "Flask-WTF",
    "flask-smorest",
    "pyyaml",
    "grpcio",
    "gunicorn",
    "gevent>=20.5.1",
    "google-api-core",
    "google-api-python-client",
    "google-auth",
    "google-cloud-core",
    # TODO(#47214): unpin bigquery when perf regression is fixed
    "google-cloud-bigquery==3.35.1",
    "google-cloud-bigquery-datatransfer",
    "google-cloud-bigquery-storage",
    "google-cloud-build",
    "google-cloud-dataflow-client",
    "google-cloud-datastore",
    "google-cloud-discoveryengine",
    "google-cloud-firestore",
    "google-cloud-logging",
    "google-cloud-monitoring",
    "google-cloud-orchestration-airflow",
    "google-cloud-run",
    "google-cloud-tasks",
    "google-cloud-pubsub",
    "google-cloud-secret-manager",
    # Version must be at least 2.18.2 to mitigate apache-beam issue
    "google-cloud-storage>=2.18.2",
    "google-cloud-trace",
    "attrs",
    "requests",
    "pytz",
    # TODO(#9732): Upgrade to 4.0 introduces dateparsing errors
    "tzlocal<4.0",
    "python-dateutil",
    "protobuf",
    "psycopg2",
    "cryptography",
    "progress",
    "pandas-gbq",
    "sendgrid",
    "psutil",
    "psycogreen",
    "marshmallow",
    "cattrs",
    # If the dill version changes be sure to update the pinned version in dataflow_setup.py
    "dill==0.3.1.1",
    # If the cloudpickle version changes be sure to update the pinned version in dataflow_setup.py
    "cloudpickle==2.2.1",
    "matplotlib",
    "scipy",
    "seaborn",
    "statsmodels",
    "openpyxl",
    "analytics-python",
    "redis",
    "pyarrow",
    # TODO(#59007): Upgrade to auth0-python>=5 - requires fixing breaking API
    # changes (auth0.exceptions and auth0.rest module paths have changed).
    "auth0-python<5",
    "sentry-sdk[flask]",
    "beautifulsoup4",
    # TODO(#46131): Unpin when sshtunnel issue is resolved
    "paramiko==3.5.1",
    "tqdm",
    "linearmodels",
    "jsonschema",
    "proto-plus",
    "nltk",
    "networkx",
    "thefuzz",
    # Needed for thefuzz to avoid "Using slow pure-python SequenceMatcher" warning
    "python-Levenshtein",
    "ratelimit",
    "pysftp",
    "PyGithub",
    "xlsxwriter",
    "pytablewriter",
    "twilio",
    "urllib3",
    "xlrd",
    "snapshottest",
    "cachetools",
    "functions-framework",
    "cloudevents",
    "quart",
    "hypercorn",
    "vertexai",
    "asyncpg",
    "pygit2",
    "matplotx",
    "sqlglot[rs]==23.11.2",
    "tabulate",
    "tenacity",
    # TODO(#46131): Unpin when sshtunnel issue is resolved
    "sshtunnel==0.4.0",
    "regex==2024.11.6",
    "pymssql",
    # llama_index and openai are tightly coupled, so we pin them together
    "llama_index>=0.13.1",
    "openai>=1.99.8",
    "litellm",
    "nest_asyncio",
]

[project.optional-dependencies]
dev = [
    "mypy-protobuf",
    "autopep8",
    "mock",
    "pytest",
    "pytest-xdist",
    "pytest-cov",
    "mypy",
    "bandit",
    "coveralls",
    # This should always be pinned to the latest version officially supported by Dataflow
    "apache-beam[gcp,test]==2.65.0",
    "astroid",
    "callee",
    "debugpy",
    "deepdiff",
    "fakeredis",
    "freezegun",
    "ipdb",
    "jupyter",
    "orjson",
    "pytest-alembic",
    "parameterized",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-genai",
    "gspread",
    # Note: Update the `black` version in `.pre-commit-config.yaml` whenever this is updated
    "black[jupyter]==22.3.0",
    "pre-commit",
    "proto-plus",
    "pygtrie",
    "rope",
    "ruamel.yaml",
    "tabula-py",
    "types-beautifulsoup4",
    "types-cachetools",
    # TODO(#60465): Un-pin when airflow environment is upgraded to a compatible version
    "types-dateparser==1.2.2.20250809",
    "types-freezegun",
    "types-grpcio",
    "types-html5lib",
    "types-httplib2",
    "types-itsdangerous",
    "types-jsonschema",
    "types-mock",
    "types-networkx",
    "types-paramiko",
    "types-protobuf",
    "types-psycopg2",
    "types-python-dateutil",
    "types-pytz",
    "types-redis",
    "types-requests",
    "types-PyYAML",
    "types-tqdm",
    "ratelimit-stubs",
    "twilio-stubs",
    "msal",
    "types-pysftp",
    "Faker",
    # TODO(https://github.com/getsentry/responses/issues/556): Unpin once responses package ships a version that doesn't break mypy
    "responses==0.20.0",
    "shellcheck-py",
    # appnope is Mac specific, included to prevent it from being dropped from lock file
    "appnope==0.1.3; sys_platform == 'darwin'",
    # recidiviz-research dependencies
    "censusdata",
    "censusgeocode",
    "Cycler",
    "fuzzywuzzy",
    "geopandas",
    "geopy",
    "h2o",
    "jupyter-client",
    "jupyter-console",
    "jupyter-core",
    "jupyterlab",
    "ipython",
    "lifelines",
    "linearmodels",
    "looker-sdk",
    "nbautoexport",
    "notebook",
    "mcp",
    "pandasql",
    "pgeocode",
    "plotly",
    "plotnine",
    "pymssql",
    "pytest_asyncio",
    "scikit-learn",
    "sentry-cli",
    "shapely<2.0.0",
    "SyntheticControlMethods",
    "ydata-profiling",
    # TODO(#21447) Unpin urllib3 when worker signal 11 failure is fixed
    "urllib3==2.6.3",
    "docker",
    "docker-stubs",
    "types-tabulate",
    "types-openpyxl",
    "ndjson",
    "nbconvert>=6.5.1",
    "xlsxwriter",
]

# Build backend for PEP 517 compliance. Hatchling is a modern, lightweight build
# backend that's simpler than setuptools. This package isn't published to PyPI,
# but uv requires a build-system for editable installs.
# See: https://hatch.pypa.io/latest/
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
force-exclude = '''
(
  /\.venv/
  | /nbautoexport-scripts/
)
'''

[tool.uv]
index-url = "https://pypi.org/simple"
# Pin setuptools < 82 for Apache Beam 2.65.0 compatibility (pkg_resources removed in 82.0.0).
# Can unpin after upgrading to apache-beam >= 2.66.0.
build-constraint-dependencies = ["setuptools<82"]
